pyproject.toml
[tool.poetry]
name = "multi-agent-etl-platform"
version = "0.1.0"
description = "A multi-agent ETL platform supporting batch and streaming data processing"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "etl_platform", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.5.0"
sqlalchemy = "^2.0.0"
kafka-python = "^2.0.2"
redis = "^5.0.0"
rq = "^1.15.0"
psycopg2-binary = "^2.9.9"
pymysql = "^1.1.0"
boto3 = "^1.34.0"
apscheduler = "^3.10.0"
pandas = "^2.1.0"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pyyaml = "^6.0.1"
langchain = "^0.1.0"
langchain-core = "^0.1.0"
langgraph = "^0.0.20"
langchain-openai = "^0.0.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
hypothesis = "^6.92.0"
pytest-asyncio = "^0.23.0"
black = "^23.12.0"
mypy = "^1.8.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py39"
